backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic[email]==2.5.2
python-multipart==0.0.6
alembic==1.12.1
ccxt==4.4.78
pandas==2.1.3
numpy==1.26.2
yfinance==0.2.32
aiohttp==3.9.1
tweepy==4.14.0
redis==5.0.1
python-whois==0.8.0
